fix DISTINCT
